+++ src/library.js
@@ -4647,18 +4647,27 @@ LibraryManager.library = {
     else return lengthBytesUTF8(str);
   },
 
+  // Returns a representation of a call site of the caller of this function, in a manner
+  // similar to __builtin_return_address. If level is 0, we return the call site of the
+  // caller of this function. This function only works when running wasm.
   emscripten_wasm_return_address__deps: ['emscripten_get_callstack_js'],
   emscripten_wasm_return_address: function(level) {
     var callstack = _emscripten_get_callstack_js(0).split('\n');
 
     // skip this function and the caller to get caller's return address
     var frame = callstack[level + 2];
-    var match = /\s+at.*wasm-function\[(\d+)\]:(\d+)/.exec(frame);
+    var match;
 
-    if (match == null) {
-      return 0;
+    if (match = /\s+at.*wasm-function\[\d+\]:(0x[0-9a-f]+)/.exec(frame)) {
+      // some engines give the binary offset directly, so we use that as return address
+      return +match[1];
+    } else if (match = /\s+at.*wasm-function\[(\d+)\]:(\d+)/.exec(frame)) {
+      // other engines only give function index and offset in the function,
+      // so we pack these into a "return address"
+      return (+match[1] << 16) + +match[2];
     } else {
-      return +match[1] << 16 | +match[2];
+      // return 0 if we can't find any
+      return 0;
     }
   },