+++ system/lib/compiler-rt/lib/ubsan_minimal/sanitizer_atomic.h
@@ -35,10 +35,6 @@ namespace __sanitizer {
 
 enum memory_order {
   memory_order_relaxed = 1 << 0,
-  memory_order_consume = 1 << 1,
-  memory_order_acquire = 1 << 2,
-  memory_order_release = 1 << 3,
-  memory_order_acq_rel = 1 << 4,
   memory_order_seq_cst = 1 << 5
 };
 
@@ -75,38 +71,16 @@ struct atomic_uintptr_t {
 
 typedef std::uintptr_t uptr;
 
-INLINE void proc_yield(int cnt) {
-  __asm__ __volatile__("" ::: "memory");
-}
-
 template<typename T>
 INLINE typename T::Type atomic_load(
     const volatile T *a, memory_order mo) {
-  DCHECK(mo & (memory_order_relaxed | memory_order_consume
-      | memory_order_acquire | memory_order_seq_cst));
+  DCHECK(mo & memory_order_relaxed);
   DCHECK(!((uptr)a % sizeof(*a)));
   typename T::Type v;
 
   if (sizeof(*a) < 8 || sizeof(void*) == 8) {
     // Assume that aligned loads are atomic.
-    if (mo == memory_order_relaxed) {
-      v = a->val_dont_use;
-    } else if (mo == memory_order_consume) {
-      // Assume that processor respects data dependencies
-      // (and that compiler won't break them).
-      __asm__ __volatile__("" ::: "memory");
-      v = a->val_dont_use;
-      __asm__ __volatile__("" ::: "memory");
-    } else if (mo == memory_order_acquire) {
-      __asm__ __volatile__("" ::: "memory");
-      v = a->val_dont_use;
-      __sync_synchronize();
-    } else {  // seq_cst
-      // E.g. on POWER we need a hw fence even before the store.
-      __sync_synchronize();
-      v = a->val_dont_use;
-      __sync_synchronize();
-    }
+    v = a->val_dont_use;
   } else {
     // 64-bit load on 32-bit platform.
     // Gross, but simple and reliable.
@@ -119,23 +93,12 @@ INLINE typename T::Type atomic_load(
 
 template<typename T>
 INLINE void atomic_store(volatile T *a, typename T::Type v, memory_order mo) {
-  DCHECK(mo & (memory_order_relaxed | memory_order_release
-      | memory_order_seq_cst));
+  DCHECK(mo & memory_order_relaxed);
   DCHECK(!((uptr)a % sizeof(*a)));
 
   if (sizeof(*a) < 8 || sizeof(void*) == 8) {
     // Assume that aligned loads are atomic.
-    if (mo == memory_order_relaxed) {
-      a->val_dont_use = v;
-    } else if (mo == memory_order_release) {
-      __sync_synchronize();
-      a->val_dont_use = v;
-      __asm__ __volatile__("" ::: "memory");
-    } else {  // seq_cst
-      __sync_synchronize();
-      a->val_dont_use = v;
-      __sync_synchronize();
-    }
+    a->val_dont_use = v;
   } else {
     // 64-bit store on 32-bit platform.
     // Gross, but simple and reliable.
@@ -162,42 +125,6 @@ INLINE void atomic_store(volatile T *a, typename T::Type v, memory_order mo) {
 // See http://www.cl.cam.ac.uk/~pes20/cpp/cpp0xmappings.html
 // for mappings of the memory model to different processors.
 
-INLINE void atomic_signal_fence(memory_order) {
-  __asm__ __volatile__("" ::: "memory");
-}
-
-INLINE void atomic_thread_fence(memory_order) {
-  __sync_synchronize();
-}
-
-template<typename T>
-INLINE typename T::Type atomic_fetch_add(volatile T *a,
-    typename T::Type v, memory_order mo) {
-  (void)mo;
-  DCHECK(!((uptr)a % sizeof(*a)));
-  return __sync_fetch_and_add(&a->val_dont_use, v);
-}
-
-template<typename T>
-INLINE typename T::Type atomic_fetch_sub(volatile T *a,
-    typename T::Type v, memory_order mo) {
-  (void)mo;
-  DCHECK(!((uptr)a % sizeof(*a)));
-  return __sync_fetch_and_add(&a->val_dont_use, -v);
-}
-
-template<typename T>
-INLINE typename T::Type atomic_exchange(volatile T *a,
-    typename T::Type v, memory_order mo) {
-  DCHECK(!((uptr)a % sizeof(*a)));
-  if (mo & (memory_order_release | memory_order_acq_rel | memory_order_seq_cst))
-    __sync_synchronize();
-  v = __sync_lock_test_and_set(&a->val_dont_use, v);
-  if (mo == memory_order_seq_cst)
-    __sync_synchronize();
-  return v;
-}
-
 template <typename T>
 INLINE bool atomic_compare_exchange_strong(volatile T *a, typename T::Type *cmp,
                                            typename T::Type xchg,
@@ -209,14 +136,6 @@ INLINE bool atomic_compare_exchange_strong(volatile T *a, typename T::Type *cmp,
   if (prev == cmpv) return true;
   *cmp = prev;
   return false;
-}
-
-template<typename T>
-INLINE bool atomic_compare_exchange_weak(volatile T *a,
-                                         typename T::Type *cmp,
-                                         typename T::Type xchg,
-                                         memory_order mo) {
-  return atomic_compare_exchange_strong(a, cmp, xchg, mo);
 }
 
 // Clutter-reducing helpers.